trading_bot/requirements.txt
requests>=2.31.0
urllib3>=2.0.0
click>=8.1.0
python-dotenv>=1.0.0
